services/ui/requirements.txt
streamlit
requests
streamlit-drawable-canvas
Pillow
